+++ imports/startup/client/Routes.jsx
@@ -21,7 +21,7 @@ const Routes = () => (
       <Switch>
         <Route path="/signin" component={LoginPage} />
         <Route path="/enroll" component={EnrollPage} />
-        {/* <Route path="/enroll-account/:token" component={EnrollPage} /> */}
+        <Route path="/enroll-account/:token" component={EnrollPage} />
         <AppContainer />
       </Switch>
     </MuiThemeProvider>
