+++ imports/ui/components/admin/EnhancedTableHead.jsx
@@ -1,0 +1,82 @@
+import React from "react";
+import PropTypes from "prop-types";
+import TableCell from "@material-ui/core/TableCell";
+import TableHead from "@material-ui/core/TableHead";
+import TableRow from "@material-ui/core/TableRow";
+import TableSortLabel from "@material-ui/core/TableSortLabel";
+import Checkbox from "@material-ui/core/Checkbox";
+import Tooltip from "@material-ui/core/Tooltip";
+
+
+const rows = [
+  { id: "practitioner", numeric: false, disablePadding: true, label: "Practitioner" },
+  { id: "backEnd", numeric: true, disablePadding: false, label: "Back-end" },
+  { id: "frontEnd", numeric: true, disablePadding: false, label: "Front-end" },
+  { id: "data", numeric: true, disablePadding: false, label: "Data" },
+  { id: "requestDate", numeric: true, disablePadding: false, label: "Request Date" }
+  // { id: 'Notes', numeric: true, disablePadding: false, label: 'Notes' },
+];
+
+export default class EnhancedTableHead extends React.Component {
+
+  createSortHandler = property => event => {
+    this.props.onRequestSort(event, property);
+  };
+
+  render() {
+    const {
+      onSelectAllClick,
+      order,
+      orderBy,
+      numSelected,
+      rowCount
+    } = this.props;
+
+    return (
+      <TableHead>
+        <TableRow>
+          <TableCell padding="checkbox">
+            <Checkbox
+              indeterminate={numSelected > 0 && numSelected < rowCount}
+              checked={numSelected === rowCount}
+              onChange={onSelectAllClick}
+            />
+          </TableCell>
+          {rows.map(
+            row => (
+              <TableCell
+                key={row.id}
+                align={row.numeric ? "right" : "left"}
+                padding={row.disablePadding ? "none" : "default"}
+                sortDirection={orderBy === row.id ? order : false}
+              >
+                <Tooltip
+                  title="Sort"
+                  placement={row.numeric ? "bottom-end" : "bottom-start"}
+                  enterDelay={300}
+                >
+                  <TableSortLabel
+                    active={orderBy === row.id}
+                    direction={order}
+                    onClick={this.createSortHandler(row.id)}
+                  >
+                    {row.label}
+                  </TableSortLabel>
+                </Tooltip>
+              </TableCell>
+            )
+          )}
+        </TableRow>
+      </TableHead>
+    );
+  }
+}
+
+EnhancedTableHead.propTypes = {
+  numSelected: PropTypes.number.isRequired,
+  onRequestSort: PropTypes.func.isRequired,
+  onSelectAllClick: PropTypes.func.isRequired,
+  order: PropTypes.string.isRequired,
+  orderBy: PropTypes.string.isRequired,
+  rowCount: PropTypes.number.isRequired
+};
